Defensive copy of invalid input in CalendricalParseExcepion.
A CharSequence (or a String) is potentially a very large object of which
only a subsequence was being parsed. Some CharSequence implementations
are mutable. An exception has a potentially unpredictable life time.
Rather than keeping a hard reference to the character sequence where the
parser error occured, make a defensive copy on construction.