Remove DateProvider, TimeProvider, DateTimeProvider
These interfaces produce unreliable APIs. For example, passing a ZonedDateTime into a DateProvider API method drops the time, offset and zone. This could cause code to do something very different to what it looks like it is doing. Removing the interfaces simplifies a lot of the API design at the expense of slightly more complictaed calls. The from(Calendrical...) methods also replace much of the use case for the providers.